+++ app/src/main/java/com/pict/pbl/medswift/screens/user_auth/RegisterScreen.kt
@@ -23,6 +23,10 @@ import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.pict.pbl.medswift.auth.CurrentUserDetails
 import com.pict.pbl.medswift.data.User
+import com.pict.pbl.medswift.screens.DatePicker
+import com.pict.pbl.medswift.screens.DropDownInput
+import com.pict.pbl.medswift.screens.ScreenTitle
+import com.pict.pbl.medswift.screens.TextInput
 import com.pict.pbl.medswift.ui.theme.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers